requirements.txt
# NOTE: Install PyTorch separately with CUDA 12.1 BEFORE installing this file:
# pip install torch==2.4.1 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

transformers==4.44.2
peft==0.12.0
bitsandbytes==0.43.3
accelerate==0.34.2
datasets==2.20.0
tokenizers==0.19.1
huggingface_hub==0.24.6
wandb==0.17.8
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
gradio==4.42.0
pyyaml==6.0.2
pytest==8.3.2
pytest-cov==5.0.0
ruff==0.6.2
black==24.8.0
